--- spec/bambora/adapters/legacy_payment_profile_response_spec.rb
@@ -1,26 +1,0 @@
-# frozen_string_literal: true
-
-require 'spec_helper'
-
-module Bambora
-  describe LegacyPaymentProfileResponse do
-    describe '#to_h' do
-      context 'a valid response' do
-        let(:json_response_body) { { someGelflings1: %w[rian deet brea] } }
-        let(:query_string_response_body) { URI.encode_www_form(json_response_body) }
-        let(:faraday_response) { Faraday::Response.new(body: query_string_response_body) }
-        subject { described_class.new(faraday_response).to_h }
-        it { is_expected.to eq some_gelflings_1: %w[rian deet brea] }
-      end
-
-      context 'an invalid response' do
-        let(:string_response_body) { 'GARTHIM! ATTACK!' }
-        let(:status) { 500 }
-        let(:faraday_response) { Faraday::Response.new(body: string_response_body, status: status) }
-        let(:error_response_hash) { { status: status, body: string_response_body } }
-        subject { described_class.new(faraday_response).to_h }
-        it { is_expected.to eq error_response_hash }
-      end
-    end
-  end
-end
