+++ spec/bambora/factories/response_adapter_factory_spec.rb
@@ -16,7 +16,7 @@ module Bambora
 
       context 'with a Query String Response' do
         let(:content_type) { 'text/html' }
-        it { is_expected.to be_a LegacyPaymentProfileResponse }
+        it { is_expected.to be_a Bank::Adapters::PaymentProfileResponse }
       end
 
       context 'with an unknown content type' do
